tensor type

+++ dataloader_2.py
@@ -93,7 +93,7 @@ class LabelDataset(Dataset):
                                    padding="max_length",
                                    add_special_tokens=True,)
             edge_index, x = self.process_graph(raw_path)
-            y = self.labels[j]
+            y = torch.LongTensor(self.labels[j])
             data = Data(x=x, y=y, edge_index=edge_index, input_ids=text_input['input_ids'], attention_mask=text_input['attention_mask'])
 
             torch.save(data, osp.join(self.processed_dir, 'data_{}.pt'.format(cid)))
